Update on the Jupiter Notebook
The notebook is complete for this simple task and commented in order to make a new user understand faster. The code is now ready to be played around with and see what changes on different parts of the code will lead to. 

Note that you have to run the anaconda notebook in the environment that Pytorch is installed in.